added ecs threat enrichment

+++ truncated-buffer_test.go
@@ -11,10 +11,10 @@ import (
 func TestTruncatedBuffer(t *testing.T) {
 	in := "12345678901234567890"
 	limit := 6
-	buff := newTruncatedBuffer(int64(limit))
+	buff := newTruncatedBuffer(limit)
 	i, err := io.Copy(buff, strings.NewReader(in))
 	assert.NoError(t, err, "should not return error")
 	assert.Equal(t, int64(len(in)), i, "should signal that it writes all byte")
-	assert.Equal(t, int64(len(in)), buff.Len(), "should return the original number of bytes to be written")
+	assert.Equal(t, len(in), buff.Len(), "should return the original number of bytes to be written")
 	assert.Equal(t, in[:limit], buff.String(), "should only contain truncated byte")
 }
